2022-10-25 Lambda 修改系统通用设置的问题

+++ erp-model/erp-model-plm/src/main/java/com/erp/model/plm/dto/ProductShowDTO.java
@@ -104,12 +104,12 @@ public class ProductShowDTO implements Serializable {
     /**
      * 立项进度
      */
-    private Integer approvalProgress;
+    private double approvalProgress;
 
     /**
      * 项目进度
      */
-    private Integer projectProgress;
+    private double projectProgress;
 
     /**
      * 任务总数
